pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "chat-gpt-local"
version = "0.1.0"
description = "Una aplicación de escritorio estilo ChatGPT integrada con OpenAI"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Developer", email = "example@example.com"}
]
dependencies = [
    "openai>=1.6.0",
    "PyQt6>=6.5.0", 
    "rich>=13.3.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.4.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.1.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "pytest>=7.3.1",
    "pytest-cov>=4.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/chat_gpt_local"]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[project.scripts]
chat-gpt-local = "chat_gpt_local.main:main"